Use manhattandistance since it results in smaller distances
Do a check before processing node to check if it's too far away to matter. Subtract the size of the box from the distance so it checks nodes slightly farther away from necessary. This ensures we don't miss any boxes because the check removed nodes that could have resulted in a closer intersection