Trying to see if i can get more ILP and better temporal cache efficiency by batching processing rays. The idea is that rays close to each other should follow each other in their path in the BVH and thus it makes sense to move multiple rays in lock step to ensure the triangles or AABB are in the lowest level of cache.
Currently it's not faster but i also don't think i am working on large enough objects for it to make a difference. May have to try it out with larger geometry so there is more cache pressure.

Of course not sure if it makes sense. it hinges on the fact that is's memory latency limited and not straight up compute bound. it also hinges on the fact that i can't make it compute bound in the future by constructing a better BVH.